refactor utility.oy; fix tests
