+++ classification/feature_generation/fishing_feature_generation.py
@@ -3,7 +3,6 @@ import numpy as np
 import tensorflow as tf
 from . import feature_generation
 
-
 def input_fn(vessel_metadata,
              filenames,
             num_features,
